test: bump series and pin revision for local charm refresh tests
A recent update to the 'ubuntu' charm requires the use of the v2
metadata format when refreshing from a local charm. The v2 format
replaces series with bases, among other changes. Properly supporting
this will need to wait on updating python-libjuju to handle bases as
well as (or instead of) series -- see issue #1156.

In the meantime, fix the test breakage by pinning the install of
'ubuntu' in broken tests to the previous revision (24). This revision
has a mismatch with the specification of the "focal" series (20.04), so
also bump the series in the local charm to "jammy".