fix: don't use lexical sorting for version numbers in codegen
Previously, python-libjuju iterated over schemas keyed by their version
string (e.g. '3.1.9') using lexical sorting. For a given facade version,
a definition in a higher versioned schema was intended to overwrite any
prior definition saved (see generate_facades function in facade.py).
With lexical sorting, '3.1.10' would be sorted in between '3.1.1' and
'3.1.2', which would not lead to the desired behaviour. This commit
fixes this problem by using a tuple of integers as the sorting key. A
special case is requried for the version string 'latest', and we use
(9000, 9000, 9000).